backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
celery>=5.3.6
redis>=5.0.3
websockets>=12.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
SQLAlchemy>=2.0.28
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
PyJWT>=2.8.1
python-jose>=3.3.0
openai>=1.14.2
pandas>=2.2.1
python-docx>=1.1.0
requests>=2.31.0
tenacity>=8.2.3
PyYAML>=6.0.1
boto3>=1.34.64
pydantic-settings
